pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "meridian"
version = "0.10.0"
description = "Offensive Recon Aggregator - passive recon across every source, simultaneously"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.47.0",
    "httpx>=0.27.0",
    "dnspython>=2.6.0",
    "python-whois>=0.9.0",
    "rich>=13.7.0",
    "click>=8.1.7",
    "python-dotenv>=1.0.0",
]

[project.scripts]
meridian = "meridian.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["meridian"]
